+++ spec/controllers/sections_controller_spec.rb
@@ -131,6 +131,15 @@ describe SectionsController do
         expect(flash[:success]).to have_message(I18n.t('sections.destroy.success'))
         expect { Section.find(section.id) }.to raise_error(ActiveRecord::RecordNotFound)
       end
+
+      it 'prevents deletion via dependent: :restrict_with_error when students exist' do
+        student = create(:student)
+        section.students << student
+
+        expect(section.destroy).to be false
+        expect(section.errors[:base]).to include('Cannot delete record because dependent students exist')
+        expect(Section.find(section.id)).to be_truthy
+      end
     end
   end
 end